requirements.txt
google-genai
pure-python-adb
pyyaml
